Forking the plugin permanently as Chainsaw and preparing for initial release
