+++ .gitignore
@@ -1,5 +1,6 @@
 .gradle
 /build/
+/out
 
 # Ignore Gradle GUI config
 gradle-app.setting
@@ -16,3 +17,4 @@ gradle-app.setting
 # IntelliJ IDEA
 .idea
 classes
+*.iml
