+++ build.gradle
@@ -1,51 +1,48 @@
 plugins {
-    id 'groovy'
-    id 'idea'
-    id 'java-gradle-plugin'
-    id 'com.gradle.plugin-publish' version '0.9.7'
+	id 'groovy'
+	id 'idea'
+	id 'java-gradle-plugin'
+	id 'com.gradle.plugin-publish' version '0.9.7'
 }
 
-group = 'org.gradle.java'
+group = 'com.zyxist'
 version = '0.1.2'
 
-targetCompatibility = '1.7'
-sourceCompatibility = '1.7'
+targetCompatibility = '1.9'
+sourceCompatibility = '1.9'
 
 // Fix a bad interaction with IntelliJ and Gradle > 4.0
 idea.module.inheritOutputDirs = true
 
-// make the publishing plugin skip checks that disallow publishing to com.gradle / org.gradle groups
-System.setProperty('gradle.publish.skip.namespace.check', 'true')
-
 repositories {
-    jcenter()
+	jcenter()
 }
 
 dependencies {
-    testCompile ('org.spockframework:spock-core:1.1-groovy-2.4') {
-        exclude module: 'groovy-all'
-    }
+	testCompile('org.spockframework:spock-core:1.1-groovy-2.4') {
+		exclude module: 'groovy-all'
+	}
 }
 
 gradlePlugin {
-    plugins {
-        experimentalJigsawPlugin {
-            id = "org.gradle.java.experimental-jigsaw"
-            implementationClass = "org.gradle.java.JigsawPlugin"
-        }
-    }
+	plugins {
+		chainsawPlugin {
+			id = "com.zyxist.chainsaw"
+			implementationClass = "com.zyxist.chainsaw.ChainsawPlugin"
+		}
+	}
 }
 
 pluginBundle {
-    website = 'https://guides.gradle.org/building-java-9-modules'
-    vcsUrl = 'https://github.com/gradle/gradle-java-modules'
-    plugins {
-        experimentalJigsawPlugin {
-            id = 'org.gradle.java.experimental-jigsaw'
-            displayName = 'Experimental Jigsaw Plugin'
-            description = 'Experiment with Java 9 modules before they are officially supported.'
-            tags = ['jigsaw', 'modules', 'java9']
-            version = project.version
-        }
-    }
+	website = 'https://github.com/zyxist/chainsaw'
+	vcsUrl = 'https://github.com/zyxist/chainsaw'
+	plugins {
+		experimentalJigsawPlugin {
+			id = 'com.zyxist.chainsaw'
+			displayName = 'Chainsaw: Java 9 Jigsaw support'
+			description = 'A plugin that adds the support for building and testing modular JDK9 applications with Gradle. '
+			tags = ['jigsaw', 'modules', 'java9']
+			version = project.version
+		}
+	}
 }
