+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = "experimental-jigsaw"
+rootProject.name = 'chainsaw'
